Dynamically add custom APNs for Sprint
For Example new devices need the new Sprint APNs, so add a new board flag for
them to set that will make the build system add the new APNs to
the APN config XML. If the flag is not set, continue to use the
old APNs.
To add or replace APNs, the custom_apns.py script has been
added. If CUSTOM_APNS_FILE is defined, custom_apns.py
is run and generates a new version of apns-conf.xml.

Also includes:

crosshatch: Add config for setting Sprint apns
	special apns such as sprint need to be added in a config file

By Josh Fox

Change-Id: I284b325524f44a1095b24c4b2be7e5b42b0026af